requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==4.2.14
django-admin-thumbnails==0.2.9
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-tinymce==4.1.0
idna==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
retrying==1.3.4
six==1.16.0
sqlparse==0.5.1
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
whitenoise==6.7.0